feat: 훅 기능 분리
