+++ src/origin/hooks/index.ts
@@ -2,6 +2,5 @@ export * from './useProduct'
 export * from './useCoupon'
 export * from './useCart'
 export * from './useContextProvider'
-export * from './useProductManager'
 export * from './useAccordion'
 export * from './useDiscount'
